Only remove very small areas from datasets with artefacts
As part of importing the geographic data we remove very small areas,
which are likely to be artefacts introduced by subtracting the sea from
the land using datasets of two separate resolutions.

We think this problem only happens with the wards and local authority
data. So we should specify that the fix should only be applied there.

This has the advantage of not removing very small areas from other
datasets, for example the Farne Islands[2] in the England polygon.

1. https://github.com/alphagov/notifications-admin/pull/3627
2. https://en.wikipedia.org/wiki/Farne_Islands