Add method to get summary of areas covered
On gov.uk/alerts we need to say roughly where an alert was sent.

There are some problems with just using the names of the areas:
- the names of electoral might not be very meaningful, for example there
  are many electoral wards called ‘Abbey’ or ‘Church’
- the names that users of the API give their areas might be very
  specific (e.g. River Coquet at Lower Stanners at Warkworth’) or not
  meant for public display (e.g. ‘Exclusion zone B’)
- the list of areas could be very long (for example with the wards
  chosen for the Reading test)

So what we need is a way to summarise the areas chosen in a way that
keeps the list short and not too specific. This way it should:
- help you disambiguate alerts on the website
- still be meaningful for people who are not very familiar with the
  area, for example those who are travelling

Using local authority names feels like the right kind of level – the
news might report something like a terror attack ‘in Birmingham’ or
flooding ‘in Cornwall’.

This commit does this by looking at the parents of the areas selected.
It adds a bit more sophistication by using the knowledge we have about
where the areas are (to calculate overlaps) and how local government is
structured (to decide whether to show the lower-tier or upper-tier name
for a local authority).