Always show form to choose / copy template
Previously a user couldn't copy an existing template if their service
only had a single notification channel enabled. While we could amend
the conditional to also show the radio buttons if the service has more
than one template, it's easier to always show it.

The only downside to always showing the form is that a user will see a
single radio button in the rare case they are adding the first template
and only one channel is enabled (for some reason).

Note that the form is currently broken if all notification channels are
disabled for a service. We don't fix that here.