Fix styling for platform admin template
The layout for the platform admin base template needed to be changed so
that the back link appears in the same place as before.

Previously, the left hand nav was inside `<main>`, but that did not need
to be and was inconsistent with other pages, so has been taken out.