Give user profile change links more context
They all currently say 'Change' which makes it
confusing when they are viewed out of their
context (ie. when all the links in the page are
listed out by a screen reader).

This gives them a suffix relating to the thing
they will change, like the links on the service
settings page.