Check that user isn’t signed in before registering
Previously this test asserted on `current_user.is_authenticated`. That
isn’t possible now because the object imported into tests isn’t the same
one the app is using.

A different proxy for whether the user is signed in is whether they have
a user id in their session, because we set this every time they sign in:
https://github.com/alphagov/notifications-admin/blob/ff32e73d9bda80e556a9ee81fb0d867eef35e1e2/app/models/user.py#L162