Make space for link in a sticky page footer
If there’s a sticky page footer which only has a right aligned link it
needs another element to give the container its height. This is because
the link itself is absolutely positioned so takes up no height.

On the _notification_ page we do this with a non-breaking space:
https://github.com/alphagov/notifications-admin/blob/1c2b65356f4b048ddea65cb5084417b2d6b68436/app/templates/views/notifications/notification.html#L101

This commit applies that same approach in the other places where we have
right-aligned links in a sticky page footer elsewhere.